requirements.txt
garminconnect>=0.2.38
gspread>=5.12.0
google-auth>=2.23.0
pytz>=2023.3
python-dotenv>=1.0.0
requests>=2.31.0
